requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
accelerate=0.27.2=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
datasets=2.17.1=pypi_0
diffusers=0.27.0.dev0=pypi_0
dill=0.3.8=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.49.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.10.0=pypi_0
ftfy=6.1.3=pypi_0
grpcio=1.62.0=pypi_0
huggingface-hub=0.20.3=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.0.1=pypi_0
jinja2=3.1.3=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown=3.5.2=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.3=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openssl=3.0.13=h7f8727e_0
packaging=23.2=pypi_0
pandas=2.2.1=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py310h06a4308_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
pyarrow=15.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyparsing=3.1.1=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
safetensors=0.4.2=pypi_0
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
torch=2.2.1=pypi_0
torchvision=0.17.1=pypi_0
tqdm=4.66.2=pypi_0
transformers=4.38.1=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py310h06a4308_0
xxhash=3.4.1=pypi_0
xz=5.4.5=h5eee18b_0
yarl=1.9.4=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
